add allergy to local storage
